Ensure itk wasm file gets built in dev
This went unnoticed before because I had run `npm run build` previously
and had the `viewer/dist` directory present. `django-vite` doesn't
forward dev server requests to the itk wasm file properly, so we just
build that file in a seperate stage in the Dockerfile and include it
in the final dev image.